+++ gdms/src/test/java/org/gdms/sql/function/ConvertFunctionTest.java
@@ -31,7 +31,7 @@
  * or contact directly:
  * info@orbisgis.org
  */
-package org.gdms.sql.function.spatial.convert;
+package org.gdms.sql.function;
 
 import com.vividsolutions.jts.geom.Coordinate;
 import com.vividsolutions.jts.geom.Geometry;
